+++ R/sbo_predictor.R
@@ -8,11 +8,12 @@ sbo_predictor <- function(object, ...)
 #' @examples
 #' \donttest{
 #' # Train a text predictor directly from corpus
-#' p <- sbo_predictor(twitter_train, N = 3, dict = max_size ~ 3,
+#' p <- sbo_predictor(twitter_train, N = 3, dict = max_size ~ 1000,
 #'                    .preprocess = preprocess, EOS = ".?!:;")
 #' }
 sbo_predictor.character <- function(object, N, dict, .preprocess = identity, 
-                                    EOS = "", lambda, L, filtered, ...) {
+                                    EOS = "", lambda = 0.4, L = 3L, 
+                                    filtered = "<UNK>", ...) {
         predtable <- sbo_predtable(object, N = N, dict = dict, 
                                    .preprocess = .preprocess, EOS = EOS,
                                    lambda = lambda, L = L, filtered = filtered)
@@ -28,7 +29,7 @@ sbo_predictor.character <- function(object, N, dict, .preprocess = identity,
 #' }
 sbo_predictor.sbo_kgram_freqs <- function(object, lambda = 0.4, L = 3L, 
                                           filtered = "<UNK>", ...) {
-        predtable <- sbo_predtable(freqs, lambda = lambda, L = L, 
+        predtable <- sbo_predtable(object, lambda = lambda, L = L, 
                                    filtered = filtered)
         return(sbo_predictor(predtable))
 }
@@ -40,18 +41,14 @@ sbo_predictor.sbo_kgram_freqs <- function(object, lambda = 0.4, L = 3L,
 #' # Load a text predictor from a Stupid Back-Off prediction table
 #' p <- sbo_predictor(twitter_predtable)
 #' }
-sbo_predictor.sbo_predtable <- function(object, ...){
-        predictor <- get_pc_ptr(object)
-        attributes(object) <- attributes(object)
-        class(object) <- c("sbo_predictor", "sbo_predictions")
-        return(object)
-}
-
-#' @rdname sbo_predictions
-#' @examples
-#' \donttest{
+#' #' \donttest{
 #' # Predict from Stupid Back-Off text predictor
 #' p <- sbo_predictor(twitter_predtable)
 #' predict(p, "i love")
 #' }
-NULL
+sbo_predictor.sbo_predtable <- function(object, ...){
+        predictor <- get_pc_ptr(object)
+        attributes(predictor) <- attributes(object)
+        class(predictor) <- c("sbo_predictor", "sbo_predictions")
+        return(predictor)
+}