+++ tests/testthat/test-print_sbo_predictor.R
@@ -1,6 +1,6 @@
 context("print.sbo_predictor")
 
-p <- load_predictor(twitter_predtable)
+p <- sbo_predictor(twitter_predtable)
 
 test_that("returns invisibly the correct value", {
         capture_output({